Add authorization to relations service
- Add UserId to Note entity
- Update Note configurations, create migration
- Add authentication, create AuthenticationOptions, update appsettings
- Add grpc users service
- Create user requests, update AddRelation, RemoveRelation
- Change AddNote command and AddNoteApplicationEventHandler
- Add UnhandledErrorException and UserNotFoundException
- Add Authorize filter to AddRelation and RemoveRelation in RelationsController
- Update tests and seed data

+++ Np.RelationsService/Np.RelationsService.Application/Relations/RemoveRelation/RemoveRelationCommandHandler.cs
@@ -24,7 +24,12 @@ namespace Np.RelationsService.Application.Relations.RemoveRelation
 
             if (relation == null) 
             {
-                return Result.Failed(Error.Null);
+                return Result.Failed(RemoveRelationErrors.NotFound);
+            }
+
+            if (!relation.Incoming.UserId.Equals(request.UserId))
+            {
+                return Result.Failed(RemoveRelationErrors.NotFound);
             }
 
             _relationsRepository.RemoveRelation(relation);
